requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albumentations==1.4.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
archspec @ file:///croot/archspec_1709217642129/work
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi==2024.8.30
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer==3.4.0
click==8.1.7
clip==0.2.0
colorama==0.4.6
comm==0.2.2
conda @ file:///croot/conda_1722004606466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.7
decorator==5.1.1
dill==0.3.8
distro @ file:///croot/distro_1714488253808/work
docker-pycreds==0.4.0
einops==0.4.0
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==4.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gpustat==1.1.1
grpcio==1.68.0
grpcio-status==1.67.1
h5py==3.12.1
httplib2==0.20.4
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
inflection==0.5.1
inquirerpy==0.3.4
iopath==0.1.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
lightning-utilities==0.11.8
lmdb==1.5.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1723567589013/work
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-flamingo==0.0.2
open_clip_torch==2.29.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pfzy==0.3.4
pillow==10.2.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
Polygon3==3.0.9.1
portalocker==2.10.1
prefetch_generator==1.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_vncorenlp==0.1.4
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjnius==1.6.1
pyparsing==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-sat==1.8.dev13
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==2.15.0
regex==2024.11.6
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==13.9.4
rotary-embedding-torch==0.8.5
rsa==4.9
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchmetrics==1.5.1
torchvision==0.17.0+cu118
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
vietocr==0.3.13
vncorenlp==1.0.3
wandb==0.18.5
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.17.1
zstandard @ file:///croot/zstandard_1714677652653/work
